+++ Sources/Runner/main.swift
@@ -4,6 +4,7 @@ import RunnerLib
 
 /// Version for showing in verbose mode
 let DangerVersion = "0.8.1"
+let MinimumDangerJSVersion = "6.1.6"
 
 private func runCommand(_ command: DangerCommand, logger: Logger) throws {
     switch command {